Move implementation unit tests to angle_unittests.
Now that libANGLE is cross-platform, this should work on OSX/Linux
and Android. Includes a warning fixes for test comparisons, and fix
to our META_ASSERT wrapper macro for clang.

BUG=angleproject:773

Reviewed-on: https://chromium-review.googlesource.com/258300
Tested-by: Jamie Madill <jmadill@chromium.org>

+++ src/libGLESv2/entry_points_gles_2_0_ext.cpp
@@ -262,7 +262,7 @@ void GL_APIENTRY GetFenceivNV(GLuint fence, GLenum pname, GLint *params)
 
           case GL_FENCE_CONDITION_NV:
             {
-                *params = fenceObject->getCondition();
+                *params = static_cast<GLint>(fenceObject->getCondition());
                 break;
             }
